+++ docs/source/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+ - CreateEmptyIndexModel and functionality
+
 ## 3.0.0 - 2021-03-19
 
 ### Added